requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pydantic==1.8.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.25
psycopg2==2.9.1
bs4
requests